core: reorganize source code to allow boardloader build
(after DMA changes)
