+++ core/embed/firmware/main.c
@@ -17,6 +17,8 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include STM32_HAL_H
+
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
