+++ core/embed/trezorhal/supervise.h
