+++ core/embed/trezorhal/systick.c
@@ -45,8 +45,6 @@
  * THE SOFTWARE.
  */
 
-#include "py/runtime.h"
-#include "py/mphal.h"
 #include "irq.h"
 #include "systick.h"
 
@@ -63,70 +61,4 @@ void SysTick_Handler(void) {
   if (f != NULL) {
     f(uw_tick);
   }
-}
-
-// We provide our own version of HAL_Delay that calls __WFI while waiting,
-// and works when interrupts are disabled.  This function is intended to be
-// used only by the ST HAL functions.
-void HAL_Delay(uint32_t Delay) {
-    if (query_irq() == IRQ_STATE_ENABLED) {
-        // IRQs enabled, so can use systick counter to do the delay
-        uint32_t start = uwTick;
-        // Wraparound of tick is taken care of by 2's complement arithmetic.
-        while (uwTick - start < Delay) {
-            // Enter sleep mode, waiting for (at least) the SysTick interrupt.
-            __WFI();
-        }
-    } else {
-        // IRQs disabled, use mp_hal_delay_ms routine.
-        mp_hal_delay_ms(Delay);
-    }
-}
-
-// Core delay function that does an efficient sleep and may switch thread context.
-// If IRQs are enabled then we must have the GIL.
-void mp_hal_delay_ms(mp_uint_t Delay) {
-    if (query_irq() == IRQ_STATE_ENABLED) {
-        // IRQs enabled, so can use systick counter to do the delay
-        uint32_t start = uwTick;
-        // Wraparound of tick is taken care of by 2's complement arithmetic.
-        while (uwTick - start < Delay) {
-            // This macro will execute the necessary idle behaviour.  It may
-            // raise an exception, switch threads or enter sleep mode (waiting for
-            // (at least) the SysTick interrupt).
-            MICROPY_EVENT_POLL_HOOK
-        }
-    } else {
-        // IRQs disabled, so need to use a busy loop for the delay.
-        // To prevent possible overflow of the counter we use a double loop.
-        const uint32_t count_1ms = HAL_RCC_GetSysClockFreq() / 4000;
-        for (int i = 0; i < Delay; i++) {
-            for (uint32_t count = 0; ++count <= count_1ms;) {
-            }
-        }
-    }
-}
-
-// delay for given number of microseconds
-void mp_hal_delay_us(mp_uint_t usec) {
-    if (query_irq() == IRQ_STATE_ENABLED) {
-        // IRQs enabled, so can use systick counter to do the delay
-        uint32_t start = mp_hal_ticks_us();
-        while (mp_hal_ticks_us() - start < usec) {
-        }
-    } else {
-        // IRQs disabled, so need to use a busy loop for the delay
-        // sys freq is always a multiple of 2MHz, so division here won't lose precision
-        const uint32_t ucount = HAL_RCC_GetSysClockFreq() / 2000000 * usec / 2;
-        for (uint32_t count = 0; ++count <= ucount;) {
-        }
-    }
-}
-
-mp_uint_t mp_hal_ticks_ms(void) {
-    return uwTick;
-}
-
-mp_uint_t mp_hal_ticks_us(void) {
-    return uwTick * 1000;
 }
